Split Tournament Home and Tournament List into separate ctrl and view pieces. Weird bug appeared whereby if headerWidget is not imported into tournamentListView, a strange second icon appears, even though headerWidget is never directly used. Importing headerWidget for now, with plans to go back and figure out what's goin on
