+++ project/src/js/ui/tournament/list/index.js
@@ -1,0 +1,7 @@
+import controller from './tournamentListCtrl';
+import view from './tournamentListView';
+
+export default {
+  controller,
+  view
+};
